implement updating comment functionality
- prevent other users from deleting comments that are not their own (admins can delete any comment)

+++ controller/comment.js
@@ -1,6 +1,7 @@
 const Comment = require("../models/comment");
 const asyncHandler = require("express-async-handler");
 const { body, validationResult } = require("express-validator");
+const comment = require("../models/comment");
 
 exports.get_all_comments = asyncHandler(async (req, res, next) => {
   const comments = await Comment.find().exec();
@@ -49,7 +50,7 @@ exports.create_comment = [
 
     await comment.save();
 
-    return res.json(comment);
+    return res.json({message: "Comment created", comment: comment});
   }),
 ];
 
@@ -61,7 +62,59 @@ exports.delete_comment = asyncHandler(async (req, res, next) => {
     return res.json(err);
   }
 
+  const comment = await Comment.findById(req.params.id).exec();
+
+  if (
+    comment.author.toString() !== req.user._id.toString() &&
+    !req.user.admin
+  ) {
+    return res
+      .status(403)
+      .json({ message: "You are not authorized to delete this comment." });
+  }
+
   await Comment.findByIdAndDelete(req.body.userId);
 
-  return res.status(200).json({ message: "Comment deleted successfully!" })
+  return res.status(200).json({ message: "Comment deleted successfully!" });
 });
+
+exports.update_comment = [
+  body("authorId").notEmpty(),
+
+  body("comment")
+    .trim()
+    .escape()
+    .isLength({ min: 1 })
+    .withMessage("Comment must have at least one character"),
+
+  asyncHandler(async (req, res, next) => {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      const errorMessages = errors.array().map((error) => ({
+        message: error.msg,
+      }));
+      return res.status(400).json({ errors: errorMessages });
+    }
+
+    const comment = await Comment.findById(req.params.id).exec();
+
+    if (comment === null) {
+      const err = new Error();
+      err.status = 404;
+      err.message = "Comment not found!";
+      return res.json(err);
+    }
+
+    const updatedComment = new Comment({
+      author: comment.author,
+      createdAt: Date.now(),
+      comment: req.body.comment,
+      post_id: comment.post_id,
+    });
+
+    await updatedComment.save();
+
+    return res.json({message: 'Your comment was updated!', comment: updatedComment});
+  }),
+];
